modularise selectVersion, update tests
