+++ packages/expo-cli/lib/test/fixtures/blank-00/package.json
@@ -1,5 +1,8 @@
 {
   "name": "bugsnag-test-fixture-04",
   "private": "true",
-  "version": "0.0.0"
+  "version": "0.0.0",
+  "dependencies": {
+    "expo": "^44.0.0"
+  }
 }
